refactor: 가독성을 높이고 모듈간 역할을 쉽게 알 수 있게 하기 위해 디렉토리 구조 변경
init/event, init/manager-page 폴더 생성
