+++ src/init/initializeSite.js
@@ -1,11 +1,7 @@
-import addMenuButton from './addMenuButton.js';
-import addMenuButtonEvent from './addMenuButtonEvent.js';
-import addSubwayMangagementContainer from './addSubwayMangagementContainer.js';
 import initPages from './page/initPages.js';
+import addMenuButtonEvent from './event/addMenuButtonEvent.js';
 
 export default function initializeSite() {
-  addMenuButton();
-  addSubwayMangagementContainer();
   initPages();
 
   addMenuButtonEvent();
