+++ src/init/page/manager-page/initSectionManagerPage.js
@@ -3,7 +3,7 @@ function sectionManagerPage() {
 }
 
 export default function initSectionManagerPage() {
-  const $managementContainer = document.querySelector('.management-container');
+  const $managementContainer = document.querySelector('.manager-container');
 
   $managementContainer.insertAdjacentHTML('beforeend', sectionManagerPage());
 }
